backend/requirements.txt
flask
flask_cors
openai
python-dotenv
pydantic
replicate
requests
azure-storage-blob
moviepy==1.0.3
Pillow==9.5.0
celery
redis
psycopg2-binary
Flask-SQLAlchemy
flask-migrate